+++ operators/src/adapters/raster_subquery/raster_subquery_adapter.rs
@@ -651,6 +651,7 @@ mod tests {
                     measurement: Measurement::Unitless,
                     time: None,
                     bbox: None,
+                    resolution: None,
                 },
             },
         }