+++ operators/src/plot/box_plot.rs
@@ -907,6 +907,7 @@ mod tests {
                         measurement: Measurement::Unitless,
                         time: None,
                         bbox: None,
+                        resolution: None,
                     },
                 },
             }
@@ -974,6 +975,7 @@ mod tests {
                         measurement: Measurement::Unitless,
                         time: None,
                         bbox: None,
+                        resolution: None,
                     },
                 },
             }
@@ -1041,6 +1043,7 @@ mod tests {
                         measurement: Measurement::Unitless,
                         time: None,
                         bbox: None,
+                        resolution: None,
                     },
                 },
             }
@@ -1107,6 +1110,7 @@ mod tests {
                         measurement: Measurement::Unitless,
                         time: None,
                         bbox: None,
+                        resolution: None,
                     },
                 },
             }
@@ -1187,6 +1191,7 @@ mod tests {
                         measurement: Measurement::Unitless,
                         time: None,
                         bbox: None,
+                        resolution: None,
                     },
                 },
             }
@@ -1259,6 +1264,7 @@ mod tests {
                         measurement: Measurement::Unitless,
                         time: None,
                         bbox: None,
+                        resolution: None,
                     },
                 },
             }
@@ -1334,6 +1340,7 @@ mod tests {
                     measurement: Measurement::Unitless,
                     time: None,
                     bbox: None,
+                    resolution: None,
                 },
             },
         };