+++ operators/src/plot/statistics.rs
@@ -470,6 +470,7 @@ mod tests {
                     measurement: Measurement::Unitless,
                     time: None,
                     bbox: None,
+                    resolution: None,
                 },
             },
         }