+++ services/src/handlers/plots.rs
@@ -256,6 +256,7 @@ mod tests {
                     measurement: Measurement::Unitless,
                     time: None,
                     bbox: None,
+                    resolution: None,
                 },
             },
         }
@@ -296,7 +297,7 @@ mod tests {
         ];
         let req = actix_web::test::TestRequest::get()
             .uri(&format!(
-                "/plot/{}/?{}",
+                "/plot/{}?{}",
                 id,
                 &serde_urlencoded::to_string(params).unwrap()
             ))
@@ -360,7 +361,7 @@ mod tests {
         ];
         let req = actix_web::test::TestRequest::get()
             .uri(&format!(
-                "/plot/{}/?{}",
+                "/plot/{}?{}",
                 id,
                 &serde_urlencoded::to_string(params).unwrap()
             ))
@@ -481,7 +482,7 @@ mod tests {
             let req = actix_web::test::TestRequest::default()
                 .method(method)
                 .uri(&format!(
-                    "/plot/{}/?{}",
+                    "/plot/{}?{}",
                     id,
                     &serde_urlencoded::to_string(params).unwrap()
                 ))