+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added spatial resolution to the `RasterResultDescriptor` that tracks the native resolution of the raster data through raster operations.
+
+- https://github.com/geo-engine/geoengine/pull/618
+
 - Added traits and methods for updating / mapping the pixels of grid types. Also added a trait for creating grid types using a function for each pixels location.
 
   - https://github.com/geo-engine/geoengine/pull/561
@@ -40,6 +44,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - https://github.com/geo-engine/geoengine/pull/614
 
 ### Changed
+
+- The Interpolation operator's input resolution can now be set to `native` to take the best available resolution, if it is known.
+
+  - https://github.com/geo-engine/geoengine/pull/618
 
 - Removed the custom Geo Bon EBV portal handlers. Instead, the EBV hierarchy is now browsed through the layer collection API.